fix: added copyright info to all files
