+++ package.json
@@ -31,6 +31,7 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@types/lodash": "^4.17.14",
     "@types/luxon": "^3.4.2",
     "@types/node": "^20",
     "@types/react": "^19",
