+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,12 +1,28 @@
-import { SessionType } from '@/lib/SessionType';
+/*
+ * Copyright 2024 Nordeck IT + Consulting GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import { jwtDecode } from 'jwt-decode';
 import NextAuth, { AuthOptions, CookiesOptions } from 'next-auth';
 import KeycloakProvider from 'next-auth/providers/keycloak';
 import { attemptAccount } from './account';
 import { refreshAccessToken } from './refreshAccessToken';
 import { attemptSessionAccessToken } from './sessionAccessToken';
-import { SessionToken, attemptSessionToken } from './sessionToken';
+import { attemptSessionToken } from './sessionToken';
 import { attemptSignInToken } from './signInToken';
+import {SessionToken, SessionType} from "@/types/types";
 
 export const authOptions: AuthOptions = {
   providers: [
