+++ src/app/api/auth/[...nextauth]/sessionToken.ts
@@ -1,22 +1,22 @@
-import Joi from 'joi';
-import {
-  SessionAccessToken,
-  sessionAccessTokenSchema,
-} from './sessionAccessToken';
-
-/**
- * User's session token.
+/*
+ * Copyright 2024 Nordeck IT + Consulting GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-export type SessionToken = {
-  name: string;
-  email: string;
-  decoded: SessionAccessToken;
-  access_token: string;
-  id_token: string;
-  expires_at: number;
-  refresh_token: string;
-  refresh_token_expired?: boolean;
-};
+
+import Joi from 'joi';
+import {sessionAccessTokenSchema,} from './sessionAccessToken';
+import {SessionToken} from "@/types/types";
 
 const sessionTokenSchema = Joi.object<SessionToken, true>({
   name: Joi.string().required(),
