+++ src/app/api/auth/[...nextauth]/testUtils.ts
@@ -1,6 +1,23 @@
+/*
+ * Copyright 2024 Nordeck IT + Consulting GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import { Account } from './account';
 import { RefreshToken } from './refreshToken';
-import { SessionToken } from './sessionToken';
+
+import {SessionToken} from "@/types/types";
 
 /**
  * {
