+++ src/contexts/Auth/AuthProvider.tsx
@@ -1,7 +1,22 @@
+/*
+ * Copyright 2024 Nordeck IT + Consulting GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use client';
 
 import { useGetEnvironment } from '@/contexts/Auth/useGetEnvironment';
-import { SessionType, UserType } from '@/lib/SessionType';
 import { authLogin, authLogout } from '@/utils/api/requests/auth.api';
 import { CircularProgress } from '@mui/material';
 import { isEqual } from 'lodash';
@@ -16,6 +31,7 @@ import {
   useMemo,
   useState,
 } from 'react';
+import {SessionType, UserType} from "@/types/types";
 
 function isPagePublic(page: string): boolean {
   return ['/login', '/meetings/meeting/join/'].some((path) =>
@@ -32,7 +48,7 @@ export type AuthState = {
 
 export const AuthContext = createContext<AuthState | undefined>(undefined);
 
-export function AuthProvider({ children }: PropsWithChildren<{}>) {
+export function AuthProvider({ children }: PropsWithChildren<object>) {
   const pathname = usePathname();
 
   const { data: session, status } = useSession();
